twitterproducer/requirements.txt
certifi==2019.6.16
chardet==3.0.4
future==0.17.1
idna==2.8
kafka-python==1.4.6
oauthlib==3.0.2
PySocks==1.7.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.8.0
urllib3==1.25.3
